requirements.txt
asyncpg
fastapi
openai
pydantic
python-dotenv
ujson
uvicorn
web3

# code quality
black
isort
